src/agent/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
langchain-ollama>=0.0.1
langchain-huggingface>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.2
httpx>=0.24.0
python-multipart>=0.0.5
chromadb>=0.4.13
sentence-transformers>=2.2.2
unstructured>=0.10.0
pypdf>=3.15.0
markdown>=3.4.3
bs4>=0.0.1
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0  # For token counting in ConversationTokenBufferMemory
redis>=5.0.0
msgpack>=1.0.5